refactor: move tests data in a specific file and improve code reuse
