+++ server/test/test-warehouse.test.ts
@@ -1,20 +1,14 @@
 import { Service } from '../src/utils/service'
-import { RequestMessage, ResponseMessage, WarehouseServiceMessages } from '../src/utils/messages';
-import { check_service } from '../src/check-service';
-
-import express from 'express';
-import { IncomingMessage, Server, ServerResponse, createServer } from 'http';
-import WebSocket, { Server as WebSocketServer } from 'ws';
+import { ResponseMessage, WarehouseServiceMessages } from '../src/utils/messages';
 import { add, cleanCollection, closeMongoClient, DbCollections, DbNames, getCollection } from './utils/db-connection';
-import { check_order_message, coffee, createRequestMessage, createResponseMessage, milk, tea } from './utils/test-utils';
+import {
+	closeWs, createConnectionAndCall, createRequestMessage, createResponseMessage,
+	initializeServer, server, startWebsocket,
+} from './utils/test-utils';
 import { ERROR_INGREDIENT_ALREADY_EXISTS, ERROR_INGREDIENT_NOT_FOUND, ERROR_INGREDIENT_QUANTITY, OK } from './utils/api-response';
+import { coffee, milk, tea } from './utils/test-data';
 
 // milk 95, tea 0
-let m: ResponseMessage
-let ws: WebSocket;
-let wss: WebSocketServer;
-const app = express();
-let server: Server<typeof IncomingMessage, typeof ServerResponse>
 let qty = 10
 
 beforeAll(async () => {
@@ -22,15 +16,14 @@ beforeAll(async () => {
 })
 
 afterEach(() => {
-	ws.close()
+	closeWs()
 	server.close()
 })
 beforeEach(async () => {
 	await cleanCollection(DbNames.WAREHOUSE, DbCollections.WAREHOUSE)
 	await add(DbNames.WAREHOUSE, DbCollections.WAREHOUSE, JSON.stringify(milk))
 	await add(DbNames.WAREHOUSE, DbCollections.WAREHOUSE, JSON.stringify(tea))
-	server = createServer(app);
-	wss = new WebSocketServer({ server });
+	initializeServer()
 })
 
 afterAll(() => { closeMongoClient() })
@@ -138,7 +131,7 @@ test('Decrease Ingredients Quantity Test - 200', done => {
 		createResponseMessage(OK, output), done)
 })
 
-function createConnectionAndCall(requestMessage: RequestMessage, expectedResponse: ResponseMessage, callback: jest.DoneCallback) {
+/*function createConnectionAndCall(requestMessage: RequestMessage, expectedResponse: ResponseMessage, callback: jest.DoneCallback) {
 	wss.on('connection', (ws) => {
 		ws.on('error', console.error);
 
@@ -168,4 +161,4 @@ function startWebsocket(requestMessage: RequestMessage, expectedResponse: Respon
 	ws.on('open', () => {
 		ws.send(JSON.stringify(requestMessage))
 	})
-}
+}*/
